requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
gunicorn==21.2.0
APScheduler==3.10.4
flask-wtf==1.2.1
PyJWT==2.8.0
email-validator==2.1.0.post1
flask-migrate==4.0.5
flask-babel==4.0.0
python-dotenv==1.0.1
Flask-Security-Too==5.3.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Admin==1.6.1
Flask-Caching==2.1.0
Flask-DebugToolbar==0.14.1
celery==5.3.6
Flask-SocketIO==5.3.6